[NDGL-44] refactor: 모듈 name space 룰에 맞춰 패키지 구조 변경
