+++ app/src/main/java/com/yapp/ndgl/ui/NDGLApp.kt
@@ -9,14 +9,14 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.navigation3.runtime.entryProvider
 import androidx.navigation3.ui.NavDisplay
-import com.yapp.travel.helper.navigation.travelHelperEntry
-import com.yapp.home.navigation.homeEntry
+import com.yapp.ndgl.feature.travelhelper.navigation.travelHelperEntry
+import com.yapp.ndgl.feature.home.navigation.homeEntry
 import com.yapp.ndgl.navigation.BottomNavTab
-import com.yapp.navigation.Navigator
-import com.yapp.navigation.Route
-import com.yapp.navigation.rememberNavigationState
-import com.yapp.navigation.toEntries
-import com.yapp.travel.navigation.travelEntry
+import com.yapp.ndgl.navigation.Navigator
+import com.yapp.ndgl.navigation.Route
+import com.yapp.ndgl.navigation.rememberNavigationState
+import com.yapp.ndgl.navigation.toEntries
+import com.yapp.ndgl.feature.travel.navigation.travelEntry
 
 @Composable
 fun NDGLApp() {
