+++ data/core/build.gradle.kts
@@ -3,5 +3,5 @@ plugins {
 }
 
 android {
-    namespace = "com.yapp.data.core"
+    namespace = "com.yapp.ndgl.data.core"
 }
