--- feature/auth/src/main/java/com/yapp/auth/AuthScreen.kt
@@ -1,41 +1,0 @@
-package com.yapp.auth
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
-import com.yapp.auth.AuthViewModel
-
-@Composable
-fun AuthRoute(
-    viewModel: AuthViewModel = hiltViewModel(),
-) {
-
-    AuthScreen()
-}
-
-@Composable
-internal fun AuthScreen(
-) {
-    LazyColumn(
-        modifier = Modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        item {
-            Text(text = "Auth Screen")
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun AuthScreenPreview() {
-    AuthScreen()
-}
