+++ feature/home/build.gradle.kts
@@ -3,7 +3,7 @@ plugins {
 }
 
 android {
-    namespace = "com.yapp.home"
+    namespace = "com.yapp.ndgl.feature.home"
 }
 
 dependencies {
