--- feature/home/src/main/java/com/yapp/home/HomeViewModel.kt
@@ -1,10 +1,0 @@
-package com.yapp.home
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class HomeViewModel @Inject constructor() : ViewModel() {
-
-}
