+++ feature/home/src/main/java/com/yapp/ndgl/feature/home/HomeViewModel.kt
@@ -1,0 +1,10 @@
+package com.yapp.ndgl.feature.home
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class HomeViewModel @Inject constructor() : ViewModel() {
+
+}
