+++ feature/home/src/main/java/com/yapp/ndgl/feature/home/navigation/HomeEntry.kt
@@ -1,0 +1,13 @@
+package com.yapp.ndgl.feature.home.navigation
+
+import androidx.navigation3.runtime.EntryProviderScope
+import androidx.navigation3.runtime.NavKey
+import com.yapp.ndgl.feature.home.HomeRoute
+import com.yapp.ndgl.navigation.Navigator
+import com.yapp.ndgl.navigation.Route
+
+fun EntryProviderScope<NavKey>.homeEntry(navigator: Navigator) {
+    entry<Route.Home> {
+        HomeRoute()
+    }
+}
