+++ feature/travel-helper/src/main/java/com/yapp/ndgl/feature/travelhelper/TravelHelperViewModel.kt
@@ -1,0 +1,10 @@
+package com.yapp.ndgl.feature.travelhelper
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class TravelHelperViewModel @Inject constructor() : ViewModel() {
+
+}
