--- feature/travel-helper/src/main/java/com/yapp/travel/helper/TravelHelperViewModel.kt
@@ -1,10 +1,0 @@
-package com.yapp.travel.helper
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class TravelHelperViewModel @Inject constructor() : ViewModel() {
-
-}
