+++ feature/travel/src/main/java/com/yapp/ndgl/feature/travel/TravelViewModel.kt
@@ -1,0 +1,10 @@
+package com.yapp.ndgl.feature.travel
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class TravelViewModel @Inject constructor() : ViewModel() {
+
+}
