--- feature/travel/src/main/java/com/yapp/travel/travel/TravelViewModel.kt
@@ -1,10 +1,0 @@
-package com.yapp.travel.travel
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class TravelViewModel @Inject constructor() : ViewModel() {
-
-}
