--- navigation/src/main/java/com/yapp/navigation/NavigationState.kt
@@ -1,70 +1,0 @@
-package com.yapp.navigation
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.snapshots.SnapshotStateList
-import androidx.compose.runtime.toMutableStateList
-import androidx.lifecycle.viewmodel.navigation3.rememberViewModelStoreNavEntryDecorator
-import androidx.navigation3.runtime.NavBackStack
-import androidx.navigation3.runtime.NavEntry
-import androidx.navigation3.runtime.NavKey
-import androidx.navigation3.runtime.rememberDecoratedNavEntries
-import androidx.navigation3.runtime.rememberNavBackStack
-import androidx.navigation3.runtime.rememberSaveableStateHolderNavEntryDecorator
-
-@Composable
-fun rememberNavigationState(
-    startRoute: NavKey,
-    topLevelKeys: Set<NavKey>,
-): NavigationState {
-    val topLevelStack = rememberNavBackStack(startRoute)
-    val subStacks = topLevelKeys.associateWith { key -> rememberNavBackStack(key) }
-
-    return remember(startRoute, topLevelKeys) {
-        NavigationState(
-            startRoute = startRoute,
-            topLevelStack = topLevelStack,
-            subStacks = subStacks,
-        )
-    }
-}
-
-class NavigationState(
-    val startRoute: NavKey,
-    val topLevelStack: NavBackStack<NavKey>,
-    val subStacks: Map<NavKey, NavBackStack<NavKey>>,
-) {
-    val currentTopLevelKey: NavKey by derivedStateOf { topLevelStack.last() }
-
-    val topLevelKeys
-        get() = subStacks.keys
-
-    val currentSubStack: NavBackStack<NavKey>
-        get() = subStacks[currentTopLevelKey]
-            ?: error("Sub stack for $currentTopLevelKey does not exist")
-
-    val currentKey: NavKey by derivedStateOf { currentSubStack.last() }
-}
-
-@Composable
-fun NavigationState.toEntries(
-    entryProvider: (NavKey) -> NavEntry<NavKey>
-): SnapshotStateList<NavEntry<NavKey>> {
-    val decoratedEntries = subStacks.mapValues { (_, stack) ->
-        val decorators = listOf(
-            rememberSaveableStateHolderNavEntryDecorator<NavKey>(),
-            rememberViewModelStoreNavEntryDecorator<NavKey>()
-        )
-        rememberDecoratedNavEntries(
-            backStack = stack,
-            entryDecorators = decorators,
-            entryProvider = entryProvider
-        )
-    }
-
-    return topLevelStack
-        .flatMap { decoratedEntries[it] ?: emptyList() }
-        .toMutableStateList()
-}
