--- navigation/src/main/java/com/yapp/navigation/Route.kt
@@ -1,22 +1,0 @@
-package com.yapp.navigation
-
-import androidx.navigation3.runtime.NavKey
-import kotlinx.serialization.Serializable
-
-@Serializable
-sealed interface Route : NavKey {
-    @Serializable
-    data object Auth : Route
-
-    @Serializable
-    data object Home : Route
-
-    @Serializable
-    data object Travel : Route
-
-    @Serializable
-    data class TravelDetail(val travelId: String) : Route
-
-    @Serializable
-    data object TravelHelper : Route
-}
