+++ navigation/src/main/java/com/yapp/ndgl/navigation/Navigator.kt
@@ -1,0 +1,44 @@
+package com.yapp.ndgl.navigation
+
+import androidx.navigation3.runtime.NavKey
+
+class Navigator(val state: NavigationState) {
+
+    fun navigate(key: NavKey) {
+        when (key) {
+            state.currentTopLevelKey -> clearSubStack()
+            in state.topLevelKeys -> goToTopLevel(key)
+            else -> goToKey(key)
+        }
+    }
+
+    fun goBack() {
+        when (state.currentKey) {
+            state.currentTopLevelKey -> {
+                state.topLevelStack.removeLastOrNull()
+            }
+
+            else -> state.currentSubStack.removeLastOrNull()
+        }
+    }
+
+    private fun goToKey(key: NavKey) {
+        state.currentSubStack.apply {
+            remove(key)
+            add(key)
+        }
+    }
+
+    private fun goToTopLevel(key: NavKey) {
+        state.topLevelStack.apply {
+            remove(key)
+            add(key)
+        }
+    }
+
+    private fun clearSubStack() {
+        state.currentSubStack.run {
+            if (size > 1) subList(1, size).clear()
+        }
+    }
+}
